requirements.txt
asteval==0.9.26
astroplan==0.8
astropy==5.3.4
asttokens==2.2.1
autograd==1.3
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
emcee==3.1.1
executing==1.2.0
extension-helpers==0.1
fonttools==4.28.3
future==0.18.2
GalSim
healpy==1.15.0
hpgeom==0.8.2
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.12.0
ipython==8.14.0
jedi==0.18.2
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lmfit==1.0.3
lsst-sphgeom==25.2023.1000
LSSTDESC.Coord==1.2.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
numpy
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
progressbar==2.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.8.1
pycparser==2.21
pyerfa==2.0.0.1
Pygments==2.15.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
scikit-learn
scipy
sgp4==2.20
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.0.0
tqdm==4.62.3
traitlets==5.9.0
typing_extensions==4.0.1
uncertainties==3.1.6
urllib3==1.26.7
wcwidth==0.2.6
ws4py==0.5.1
xfiles==0.1.0
zipp==3.6.0
llnl-ssapy